src/test/java/GerenciadorTransacaoTest.java
import org.example.wallace.GerenciadorTransacao;
import org.example.wallace.GerenciadorTransacaoNoSQL;
import org.example.wallace.GerenciadorTransacaoSQL;
import org.junit.Assert;
import org.junit.Test;

public class GerenciadorTransacaoTest {

    @Test
    public void testValidarTransacaoSQL() throws Exception {
        GerenciadorTransacao gerenciadorSQL = new GerenciadorTransacaoSQL();
        ((GerenciadorTransacaoSQL) gerenciadorSQL).validarTransacao();

        // Verifica se a validação da transação SQL foi bem-sucedida
        Assert.assertTrue(verificarValidacaoTransacaoSQL());
    }

    @Test
    public void testValidarTransacaoNoSQL() throws Exception {
        GerenciadorTransacao gerenciadorNoSQL = new GerenciadorTransacaoNoSQL();
        ((GerenciadorTransacaoNoSQL) gerenciadorNoSQL).validarTransacao();

        // Verifica se a validação da transação NoSQL foi bem-sucedida
        Assert.assertTrue(verificarValidacaoTransacaoNoSQL());
    }

    @Test
    public void testFinalizarTransacaoSQL() throws Exception {
        GerenciadorTransacao gerenciadorSQL = new GerenciadorTransacaoSQL();
        ((GerenciadorTransacaoSQL) gerenciadorSQL).finalizarTransacao();

        // Verifica se a finalização da transação SQL foi bem-sucedida
        Assert.assertTrue(verificarFinalizacaoTransacaoSQL());
    }

    @Test
    public void testFinalizarTransacaoNoSQL() throws Exception {
        GerenciadorTransacao gerenciadorNoSQL = new GerenciadorTransacaoNoSQL();
        ((GerenciadorTransacaoNoSQL) gerenciadorNoSQL).finalizarTransacao();

        // Verifica se a finalização da transação NoSQL foi bem-sucedida
        Assert.assertTrue(verificarFinalizacaoTransacaoNoSQL());
    }

    private boolean verificarValidacaoTransacaoSQL() {
        // Lógica para verificar se a validação da transação SQL foi bem-sucedida
        // Retorne true se a validação estiver correta, ou false caso contrário
        return true;
    }

    private boolean verificarValidacaoTransacaoNoSQL() {
        // Lógica para verificar se a validação da transação NoSQL foi bem-sucedida
        // Retorne true se a validação estiver correta, ou false caso contrário
        return true;
    }

    private boolean verificarFinalizacaoTransacaoSQL() {
        // Lógica para verificar se a finalização da transação SQL foi bem-sucedida
        // Retorne true se a finalização estiver correta, ou false caso contrário
        return true;
    }

    private boolean verificarFinalizacaoTransacaoNoSQL() {
        // Lógica para verificar se a finalização da transação NoSQL foi bem-sucedida
        // Retorne true se a finalização estiver correta, ou false caso contrário
        return true;
    }
}